Feat: Add MaintainedAdmonition component
Adds MaintainedAdmonition component to allow configuring a conditional render which warns users that a particular course may be unmaintained and out-of-date.

This can be set in the `config.js` for any course using `siteMetadata.isMaintained`; setting it to false will render the admonition while setting it to true — or omitting it — will result in the component not being visible.